remove broken conditional require
`view' was always required because it is not only required within the
`if' but also outside.  Therefore no body used this using an Emacsen
that would have caused `view-less' to be loaded "instead", otherwise
they would have reported a bug.  So we can drop the `if` completely
and stick to the unconditional (require 'view).