fixup copyright notices, without changing license
This commit merely changes the formatting and/or wording of the
copyright notices at the top of the individual libraries.  The
used license is not changed, it was and still is the GNU GPL v2,
with the "(at your option) any later version" note.

The formatting is changed to follow the formatting that is commonly
used nowadays.  This helps tools that try to extract the license based
on the license statement.

`debian-el.el' previously had no license statement, but because all
the other license statements revered to the same license, it is clear
that this file was and has to be released under the same license.
This is the file where the license statement is most important,
because it is the "main library", where tools most likely look for it.